+++ app/src/frontend/config/category-ui-config.ts
@@ -1,16 +1,16 @@
 import { Category } from './categories-config';
 
-import AgeContainer from '../building/data-containers/age-history';
+import PlanningConservationContainer from '../building/data-containers/planning-construction';
 import CommunityContainer from '../building/data-containers/community';
-import ConstructionContainer from '../building/data-containers/construction';
-import ResilienceContainer from '../building/data-containers/resilience';
+import AgeHistoryContainer from '../building/data-containers/age-history';
+import DisasterManagementContainer from '../building/data-containers/disaster-management';
 import LocationContainer from '../building/data-containers/location';
-import PlanningContainer from '../building/data-containers/planning';
-import SizeContainer from '../building/data-containers/size';
-import StreetContextContainer from '../building/data-containers/street-context';
-import SustainabilityContainer from '../building/data-containers/energy-performance';
-import TeamContainer from '../building/data-containers/team';
-import TypeContainer from '../building/data-containers/typology';
+import UrbanInfrastructureContainer from '../building/data-containers/urban-infrastructure';
+import ConstructionDesignContainer from '../building/data-containers/construction-design';
+import RetrofitConditionContainer from '../building/data-containers/retrofit-condition';
+import WaterGreenInfrastructureContainer from '../building/data-containers/water-green-infrastructure';
+import EnergyPerformanceContainer from '../building/data-containers/energy-performance';
+import TypologySizeContainer from '../building/data-containers/typology-size';
 import UseContainer from '../building/data-containers/land-use';
 
 import { DataContainerType } from '../building/data-container';
@@ -18,15 +18,15 @@ import { DataContainerType } from '../building/data-container';
 export const categoryUiConfig: {[key in Category]: DataContainerType} = {
     [Category.Location]: LocationContainer,
     [Category.LandUse]: UseContainer,
-    [Category.TypologySize]: TypeContainer,
-    [Category.AgeHistory]: AgeContainer,
-    [Category.ConstructionDesign]: SizeContainer,
-    [Category.PlanningConservation]: ConstructionContainer,
-    [Category.RetrofitCondition]: StreetContextContainer,
-    [Category.EnergyPerformance]: TeamContainer,
-    [Category.UrbanInfrastructure]: PlanningContainer,
-    [Category.WaterGreenInfrastructure]: SustainabilityContainer,
-    [Category.DisasterManagement]: ResilienceContainer,
+    [Category.TypologySize]: TypologySizeContainer,
+    [Category.AgeHistory]: PlanningConservationContainer,
+    [Category.ConstructionDesign]: ConstructionDesignContainer,
+    [Category.PlanningConservation]: AgeHistoryContainer,
+    [Category.RetrofitCondition]: RetrofitConditionContainer,
+    [Category.EnergyPerformance]: EnergyPerformanceContainer,
+    [Category.UrbanInfrastructure]: UrbanInfrastructureContainer,
+    [Category.WaterGreenInfrastructure]: WaterGreenInfrastructureContainer,
+    [Category.DisasterManagement]: DisasterManagementContainer,
     [Category.Community]: CommunityContainer,
 };
